feat: add dokka documentation
