+++ src/main/kotlin/io/sos/alisos/Application.kt
@@ -2,6 +2,9 @@ package io.sos.alisos
 
 import io.micronaut.runtime.Micronaut
 
+/**
+ * Main object for run Application
+ */
 object Application {
     @JvmStatic
     fun main(args: Array<String>) {
