+++ src/main/kotlin/io/sos/alisos/controller/AliceController.kt
@@ -9,6 +9,11 @@ import io.sos.alisos.service.AliceService
 import org.slf4j.LoggerFactory
 import javax.inject.Inject
 
+/**
+ * Controller for http request of our service
+ *
+ * @property service service with the logic of processing a received request
+ */
 @Controller
 class AliceController {
     @Inject
@@ -16,11 +21,16 @@ class AliceController {
 
     var logger = LoggerFactory.getLogger(AliceController::class.java)
 
+    /**
+     * request handler from Yandex Alice
+     */
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     @Post("/webhook")
     fun webhook(@Body request: RequestWrapper): ResponseWrapper {
         val userId = request.session.userId
+
+        logger.info("webhook requested: userId: {}", userId)
 
         val user = MessageInfo(
             anamnesis = request.request.originalUtterance,
@@ -32,10 +42,16 @@ class AliceController {
             no = request.request.nlu?.tokens?.any { it.equals("нет", true) } ?: false
         )
 
+        logger.debug("webhook requested: userId: {}, user information: {}", userId, user)
+
+        val response = service.webhook(userId, user)
+
+        logger.debug("webhook service return: userId: {}, response: {}", userId, response)
+
         return ResponseWrapper(
             session = request.session,
             version = "1.0",
-            response = service.webhook(userId, user)
+            response = response
         )
     }
 }
