+++ src/main/kotlin/io/sos/alisos/domain/RequestDomain.kt
@@ -7,6 +7,13 @@ import io.micronaut.core.annotation.Introspected
 import java.math.BigDecimal
 import javax.validation.constraints.NotBlank
 
+/**
+ * User session
+ *
+ * @property sessionId Unique session identifier, maximum 64 characters.
+ * @property messageId Session message identifier, maximum 8 characters.
+ * @property userId The identifier of the application instance in which the user communicates with Alice is a maximum of 64 characters.
+ */
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
 @Introspected
 open class Session {
@@ -20,6 +27,9 @@ open class Session {
     lateinit var userId: String
 }
 
+/**
+ * Location recognized by Alice from the user's message
+ */
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
 @Introspected
 data class YandexGeo(
@@ -30,6 +40,9 @@ data class YandexGeo(
     var airport: String?
 )
 
+/**
+ * Fio recognized by Alice from the user's message
+ */
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
 @Introspected
 data class YandexFio(
@@ -38,6 +51,9 @@ data class YandexFio(
     var patronymicName: String?
 )
 
+/**
+ * Date recognized by Alice from the user's message
+ */
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
 @Introspected
 data class YandexDatetime(
@@ -53,6 +69,9 @@ data class YandexDatetime(
     var minuteIsRelative: Boolean?
 )
 
+/**
+ * Entity recognized by Alice from the user's message
+ */
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
 @JsonDeserialize(using = EntitiesDeserializer::class)
 @Introspected
@@ -64,14 +83,21 @@ data class Entity(
     var yandexDatetime: YandexDatetime? = null
 )
 
+/**
+ * Words and named entities that Alice have extracted from a user request.
+ */
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
 @Introspected
 open class NLU {
     lateinit var tokens: Array<String>
     lateinit var entities: Array<Entity>
-
 }
 
+/**
+ * Data received from the user.
+ *
+ * @property originalUtterance full text of user request, maximum 1024 characters.
+ */
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
 @Introspected
 open class Request {
@@ -81,6 +107,9 @@ open class Request {
     var nlu: NLU? = null
 }
 
+/**
+ * Data received from Alice.
+ */
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy::class)
 @Introspected
 open class RequestWrapper {
